% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composeTable.R
\name{composeTable}
\alias{composeTable}
\title{Compose package tables from component data}
\usage{
composeTable(
  name,
  col_names,
  row_names,
  merge,
  merge_dir = NULL,
  data_dir = NULL,
  verbose = FALSE,
  merge_verbose = FALSE
)
}
\arguments{
\item{name}{Character vector of dataset name}

\item{col_names}{Character vector of columns for dataset. Set to NULL to not specify any column names.}

\item{row_names}{Character vector of rows for dataset. Set to NULL to not specify any row names.}

\item{merge}{Boolean to indicate merging of component data.}

\item{merge_dir}{Character vector of directory pathway to component data if merging is needed. If not specified, data will be accessed from designated package extdata directory. (Default: NULL)}

\item{data_dir}{Character vector of directory pathway to data file if merging is not needed. Also the destination for the data after composition and conversion. If not specified, data will be saved to designated package extdata directory. (Default: NULL)}

\item{verbose}{A boolean to print result summary for saving objects. (Default: FALSE)}

\item{merge_verbose}{A boolean to print result summary for data merges. (Default: FALSE)}
}
\description{
Composes and converts the following dataset directories/files to .RData format in post-installation
when run in \code{composeData()}:
}
\examples{
\dontrun{
composeTable(name = "GEO_genes",
             col_names = c("ensembl_id", "gene_short_name"),
             row_names = "ensembl_id",
             merge = FALSE,
             merge_dir = NULL, data_dir = NULL,
             verbose = FALSE, merge_verbose = FALSE)
}
}
