% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvMerge.R
\name{csvMerge}
\alias{csvMerge}
\title{CSV merge function for CSV component files}
\usage{
csvMerge(data_dir, axis = 0)
}
\arguments{
\item{data_dir}{A character vector with one element as the path to the directory of components to merge}

\item{axis}{A numeric vector with one element to specify axis to merge (0: rows, 1: columns)}
}
\value{
A Data.Frame dataset
}
\description{
CSV merge function for CSV component files
}
\examples{
\dontrun{
csvMerge(data_dir = "data/example_data/data")
}
}
