% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvSplit.R
\name{csvSplit}
\alias{csvSplit}
\title{CSV split function for large CSV files}
\usage{
csvSplit(data_file, nb_comp, axis = 0, verbose = FALSE)
}
\arguments{
\item{data_file}{A character vector with one element as the path to the file to split}

\item{nb_comp}{A numeric vector with one element to specify how many components to split the file into}

\item{axis}{A numeric vector with one element to specify axis to split (0: rows, 1: columns) (Default: 0)}

\item{verbose}{A boolean to print result summary (Default: FALSE)}
}
\description{
CSV split function for large CSV files
}
\examples{
\dontrun{
csvSplit(data_file = "data/example_data/data.csv", nb_comp = 2, axis = 0, verbose = FALSE)
}
}
