% Generated by authors
\docType{data}
\name{ex_data}
\alias{ex_data}
\title{Example csv dataset}
\format{
\subsection{\code{ex_data}}{

A data frame with 22 rows and 3 columns:
\describe{
\item{i}{`numeric` vector}
\item{j}{`numeric` vector}
\item{k}{`numeric` vector}
}
}
}
\source{
Randomly generated by authors
}
\usage{
ex_data
}
\description{
A dataset of randomly generated numeric values to demonstrate package functions
}
\keyword{datasets}
