% Generated by authors
\docType{data}
\name{ex_matrix}
\alias{ex_matrix}
\title{Example matrix dataset}
\format{
\subsection{\code{ex_matrix}}{

A matrix with 3 rows and 3 columns:
\describe{
\item{i}{`integer` row indices of non-zero entries in 0-base}
\item{j}{`integer` column indices of non-zero entries in 0-base}
\item{x}{`numeric` vector - the non-zero entry at position `(i, j)`}
\item{Dim}{`integer` of length 2 - the dimensions of the matrix}
}
}
}
\source{
Randomly generated by authors
}
\usage{
ex_matrix
}
\description{
A matrix of numeric values to demonstrate package functions
}
\keyword{datasets}
