% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixify.R
\name{matrixify}
\alias{matrixify}
\title{Convert partial data to a matrix with a metadata list for eventual SummarizedExperiment objects}
\usage{
matrixify(
  data,
  names = NULL,
  subs = NULL,
  transpose = FALSE,
  verbose = FALSE,
  merge_verbose = FALSE
)
}
\arguments{
\item{data}{A DataFrame object}

\item{names}{A list of length three with vectors in "col", "row", and "meta" for naming. (Default: NULL)}

\item{subs}{A list of length two with vectors in "col" and "row" for subsetting. (Default: NULL)}

\item{transpose}{A bollean to indicate if matrix should be transposed. (Default: FALSE)}

\item{verbose}{A boolean to print result summary for saving objects. (Default: FALSE)}

\item{merge_verbose}{A boolean to print result summary for data merges. (Default: FALSE)}
}
\value{
A list of 2 elements with (1) the matrix, and (2) the metadata list.
}
\description{
Convert partial data to a matrix with a metadata list for eventual SummarizedExperiment objects
}
\examples{
\dontrun{
?mpr_merfish

names = list(
  col = colnames(mpr_merfish),
  row = mpr_merfish$Cell_ID,
  meta = c("Cell_ID", "Centroid_X", "Centroid_Y")
)

subs = list(
  col = names[["col"]][grep("blank", names[["col"]], ignore.case = TRUE)],
  row = names[["row"]][1:100]
)

mpr_merfish_mat = matrixify(data = mpr_merfish,
                            names = names,
                            subs = subs,
                            transpose = TRUE)
}
}
