% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixMerge.R
\name{matrixMerge}
\alias{matrixMerge}
\title{Matrix merge function for Matrix component files}
\usage{
matrixMerge(data_dir, verbose = FALSE)
}
\arguments{
\item{data_dir}{A character vector with one element as the path to the directory of components to merge}

\item{verbose}{A boolean to print result summary}
}
\value{
A dgTMatrix object dataset
}
\description{
Matrix merge function for Matrix component files
}
\examples{
\dontrun{
matrixMerge(data_dir = "data/example_data/matrix")
}
}
