% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixSplit.R
\name{matrixSplit}
\alias{matrixSplit}
\title{Matrix split function for large Matrix files}
\usage{
matrixSplit(data_file, nb_comp, verbose = FALSE)
}
\arguments{
\item{data_file}{A character vector with one element as the path to the file to split}

\item{nb_comp}{A numeric vector with one element to specify how many components to split the file into}

\item{verbose}{A boolean to print result summary (Default: FALSE)}
}
\description{
Matrix split function for large Matrix files
}
\examples{
\dontrun{
matrixSplit(data_file = "data/example_data/matrix.mtx", nb_comp = 2, verbose = FALSE)
}
}
