% Generated by authors
\docType{data}
\name{mpr_barcodes}
\alias{mpr_barcodes}
\title{Barcodes data for GEO data}
\format{
\subsection{\code{mpr_barcodes}}{

A data frame with 7500 rows and 1 columns:
\describe{
\item{barcode}{`character` vector - cell barcodes}
}
}
}
\source{
<https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE113576>
}
\usage{
mpr_barcodes
}
\description{
A dataset of cell barcodes for the GEO data (first 7500 rows)
}
\keyword{datasets}
