% Generated by authors
\docType{data}
\name{mpr_genes}
\alias{mpr_genes}
\title{Gene ID data for GEO data}
\format{
\subsection{\code{mpr_genes}}{

A data frame with 27998 rows and 2 columns:
\describe{
\item{ensembl}{`character` vector - ENSEMBL Gene IDS}
\item{gene_short_name}{`character` vector - Short Gene Names}
}
}
}
\source{
<https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE113576>
}
\usage{
mpr_genes
}
\description{
A dataset of genes in ENSEMBL and Short Gene Name formats for the GEO data (first 7500 rows)
}
\keyword{datasets}
