% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMPR.R
\name{readMPR}
\alias{readMPR}
\title{Load in data from local directory similar to Seurat's \code{Read10X}}
\usage{
readMPR(data_dir, cell_column = 1, gene_column = 1, colnames = NULL)
}
\arguments{
\item{data_dir}{Directory containing the matrix.mtx, genes.tsv, barcodes.tsv files to load}

\item{cell_column}{Specify which column of barcodes.tsv to use for cell names. (Default: 1)}

\item{gene_column}{Specify which column of genes.tsv to use for gene names. (Default: 1)}

\item{colnames}{List of column names for barcodes and genes data sets (Default: NULL)}
}
\value{
An expression matrix with updated dimension names for genes and cells
}
\description{
Load in data from local directory similar to Seurat's \code{Read10X}
}
\examples{
\dontrun{
mpr = readMPR(data_dir = "mpr_experiment",
               cell_column = 1,
               gene_column = 2,
               colnames = list(barcodes_col = c("barcode"),
                               genes_col = c("ensembl_id", "gene_short_name")))
}
}
