% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composeData.R
\name{composeData}
\alias{composeData}
\title{Compose package datasets from component data}
\usage{
composeData(verbose = FALSE, merge_verbose = FALSE)
}
\arguments{
\item{verbose}{A boolean to print result summary for saving objects (Default: FALSE)}

\item{merge_verbose}{A boolean to print result summary for data merges (Default: FALSE)}
}
\description{
Composes and/or converts the following dataset directories/files to .RData format in post-installation
when user runs \code{composeData()}:
\itemize{
\item bio data
\itemize{
\item mpr_barcodes.tsv (tsv dataset)
\item mpr_genes.tsv (tsv dataset)
\item mpr_matrix/ (matrix components)
\item mpr_merfish/ (csv components)
}
}
}
\examples{
\dontrun{
composeData(verbose = FALSE, merge_verbose = FALSE)
}
}
