% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composeData.R
\name{composeData}
\alias{composeData}
\title{Compose package datasets from component data}
\usage{
composeData(verbose = FALSE, merge_verbose = FALSE)
}
\arguments{
\item{verbose}{A boolean to print result summary for saving objects}

\item{merge_verbose}{A boolean to print result summary for data merges}
}
\description{
Compose package datasets from component data
}
\examples{
\dontrun{
composeData()
}
}
