% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Load package datasets from composed Rdata}
\usage{
loadData(verbose = FALSE)
}
\arguments{
\item{verbose}{A boolean to print result summary for loading objects (Default: FALSE)}
}
\description{
Loads package datasets to global environment. User has the option to load all, only example, or only bio datasets.
\itemize{
\item bio data
\itemize{
\item mpr_barcodes
\item mpr_genes
\item mpr_matrix
\item mpr_merfish
}
}
}
\examples{
\dontrun{
loadData(verbose = FALSE)
}
}
