% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Load package datasets from composed Rdata}
\usage{
loadData(datasets = "all", verbose = FALSE)
}
\arguments{
\item{datasets}{Character vector to select which datasets to load - one of 'all', 'bio', or 'example' (Default: "all")}

\item{verbose}{A boolean to print result summary for loading objects (Default: FALSE)}
}
\description{
Load package datasets from composed Rdata
}
\examples{
\dontrun{
loadData(datasets = "all", verbose = FALSE)
}
}
