% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Load package datasets from composed Rdata}
\usage{
loadData(verbose = TRUE)
}
\arguments{
\item{verbose}{A boolean to print result summary for loading objects}
}
\description{
Load package datasets from composed Rdata
}
\examples{
\dontrun{
loadData()
}
}
