% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixify.R
\name{matrixify}
\alias{matrixify}
\title{Convert partial data to a matrix with a metadata list for eventual SummarizedExperiment objects}
\usage{
matrixify(
  data,
  col_names = NULL,
  row_names = NULL,
  col_subset = NULL,
  row_subset = NULL,
  col_meta = NULL,
  transpose = FALSE,
  verbose = FALSE,
  merge_verbose = FALSE
)
}
\arguments{
\item{data}{A DataFrame object}

\item{col_names}{A vector of column names or numbers. (Default: NULL)}

\item{row_names}{A vector of row names or numbers. (Default: NULL)}

\item{col_subset}{A vector of columns to subset data with. (Default: NULL)}

\item{row_subset}{A vector of rows to subset data with. (Default: NULL)}

\item{col_meta}{A vector of columns to specify as meta data columns. (Default: NULL)}

\item{transpose}{A bollean to indicate if matrix should be transposed. (Default: FALSE)}

\item{verbose}{A boolean to print result summary for saving objects. (Default: FALSE)}

\item{merge_verbose}{A boolean to print result summary for data merges. (Default: FALSE)}
}
\value{
A list of 2 elements with \link{1} the matrix, and \link{2} the metadata list.
}
\description{
Convert partial data to a matrix with a metadata list for eventual SummarizedExperiment objects
}
\examples{
\dontrun{
?DRYAD_data

col_names = colnames(DRYAD_data)
row_names = DRYAD_data$Cell_ID

col_subset = col_names[grep("blank", col_names, ignore.case = TRUE)]
row_subset = row_names[1:100]

DRYAD_matrix = matrixify(data = DRYAD_data,
                         col_names = col_names,
                         row_names = row_names,
                         col_subset = col_subset,
                         row_subset = row_subset,
                         col_meta = c("Cell_ID", "Centroid_X", "Centroid_Y"),
                         transpose = TRUE)
}
}
